flow_package_maths/arithmetic/negative/negative.py
from flow import Ports, Process
from flow_types import base, unions

# Define Ports
ports = Ports()

# Add Inports
ports.add_inport(id="value", types=unions.Number)

# Add Outports
ports.add_outport(id="result", types=[base.Double])


def process(component: Process):

    if not component.has_data():
        return

    # Get Inport Data
    value = float(component.get_data("value"))

    out_val = -value

    # Send Outport Data
    component.send_data(base.Double(out_val), "result")
