pyproject.toml
[tool.poetry]
name = "flow-package-maths"
description = "A Package of Python Flow Components"
authors = [
  "Arran Hobson Sayers <arran@theengineeringcompany.com>",
  "Eu Chian Tan <euchian@theengineeringcompany.com>",
  "Neil Cameron <neil@theengineeringcompany.com>",
  "Harry Thompson <harry@theengineeringcompany.com>",
]
version = "3.0.0-dev.6"
license = "MIT"
readme = "README.md"
repository = "https://github.com/theengineeringco/flow-package-maths.git"

[tool.poetry.dependencies]
python  = "^3.8"
flow-py = "^3-dev"
numpy   = "^1.18"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.15"  # Linting
mypy                     = "^0.812" # Linting
black                    = "^21b"   # Formatting
isort                    = "^5.0"   # Formatting
unimport                 = "^0.8"   # Formatting
pytest                   = "^6.0"   # Tests
tox                      = "^3.0"   # Tests
pre-commit               = "^2.10"  # Code Style

[[tool.poetry.source]]
name = "tec-pypi"
url  = "https://pypi.theengineering.company"

[tool.poetry.plugins.flow_package]
"maths" = "flow_package_maths"

[tool.nitpick]
style = "https://raw.githubusercontent.com/theengineeringco/python-styleguide/master/TEC-style.toml"

[tool.pytest.ini_options]
minversion     = "6.0"
testpaths      = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.black]
line-length    = 120
target-version = ["py38"]

[tool.isort]
line_length = 120
profile     = "black"

[tool.unimport]
include = '^flow_package_maths|^tests'
exclude = '(__init__.py)'

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
